backend/app/requirements.txt
asttokens==2.0.5
autopep8==1.6.0
backcall==0.2.0
bcrypt==3.2.0
black==21.12b0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cx-Oracle==8.3.0
flask-cors==4.0.0
decorator==5.1.1
executing==0.8.2
Flask==2.0.2
flask-socketio==5.3.4
eventlet==0.33.0
ipython==8.0.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
numpy==1.22.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
platformdirs==2.4.1
prompt-toolkit==3.0.24
pure-eval==0.2.2
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
requests==2.30.0
schedule==1.2.0
six==1.16.0
stack-data==0.1.4
toml==0.10.2
tomli==1.2.3
traitlets==5.1.1
typing_extensions==4.0.1
wcwidth==0.2.5
Werkzeug==2.0.2
psycopg2-binary==2.9.2
